@@ -54,8 +54,8 @@ const SERVICES = [
   ],
 })
 export class MockDataModule {
-  static forRoot(): ModuleWithProviders {
-    return <ModuleWithProviders>{
+  static forRoot(): ModuleWithProviders<MockDataModule> {
+    return {
       ngModule: MockDataModule,
       providers: [
         ...SERVICES,
