@@ -1,7 +1,5 @@
 import { Component, ElementRef, EventEmitter, NgZone, OnInit, Output, ViewChild } from '@angular/core';
-import { MapsAPILoader } from '@agm/core';
-import { Location } from '../entity/Location';
-
+import { PositionModel } from '../entity/position.model';
 
 @Component({
   selector: 'ngx-search',
@@ -9,35 +7,33 @@ import { Location } from '../entity/Location';
 })
 export class SearchComponent implements OnInit {
 
-  @Output() positionChanged = new EventEmitter<Location>();
+  @Output()
+  positionChanged: EventEmitter<PositionModel> = new EventEmitter<PositionModel>();
 
   @ViewChild('search', { static: true })
-  public searchElementRef: ElementRef;
+  searchElementRef: ElementRef;
 
-  constructor(private mapsAPILoader: MapsAPILoader,
-              private ngZone: NgZone) {
-  }
+  constructor(private ngZone: NgZone) {}
 
   ngOnInit() {
-    // load Places Autocomplete
-    this.mapsAPILoader.load().then(() => {
-      const autocomplete = new google.maps.places.Autocomplete(this.searchElementRef.nativeElement, {
-        types: ['address'],
-      });
-      autocomplete.addListener('place_changed', () => {
-        this.ngZone.run(() => {
-          // get the place result
-          const place: google.maps.places.PlaceResult = autocomplete.getPlace();
+    const autocomplete = new google.maps.places.Autocomplete(
+      this.searchElementRef.nativeElement, { types: ['address'] },
+    );
 
-          // verify result
-          if (place.geometry === undefined || place.geometry === null) {
-            return;
-          }
+    autocomplete.addListener('place_changed', () => {
+      this.ngZone.run(() => {
+        // get the place result
+        const place: google.maps.places.PlaceResult = autocomplete.getPlace();
 
-          this.positionChanged.emit(
-            new Location(place.geometry.location.lat(),
-              place.geometry.location.lng()));
-        });
+        // verify result
+        if (place.geometry === undefined || place.geometry === null) {
+          return;
+        }
+
+        this.positionChanged.emit(new PositionModel(
+          place.geometry.location.lat(),
+          place.geometry.location.lng(),
+        ));
       });
     });
   }
