@@ -1,13 +1,11 @@
 import { Component } from '@angular/core';
-import { ToasterConfig } from 'angular2-toaster';
-
-import 'style-loader!angular2-toaster/toaster.css';
 import {
   NbComponentStatus,
   NbGlobalLogicalPosition,
   NbGlobalPhysicalPosition,
   NbGlobalPosition,
   NbToastrService,
+  NbToastrConfig,
 } from '@nebular/theme';
 
 @Component({
@@ -18,7 +16,7 @@ import {
 export class ToastrComponent {
   constructor(private toastrService: NbToastrService) {}
 
-  config: ToasterConfig;
+  config: NbToastrConfig;
 
   index = 1;
   destroyByClick = true;
